# auth-service/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth-service_internal_domain.Group:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      updated_at:
        type: string
    type: object
  auth-service_internal_domain.GroupMember:
    properties:
      created_at:
        type: string
      group_id:
        type: string
      invited_by:
        type: string
      joined_at:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  auth-service_internal_domain.Permission:
    properties:
      group_id:
        type: string
      metric_type:
        type: string
      user_id:
        type: string
    type: object
...
  auth-service_internal_web_helpers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  auth-service_internal_web_helpers.OKResponse:
    properties:
      message:
        type: string
    type: object
...
  internal_group.CreateGroupRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  internal_group.InviteMemberRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_group.SetPermissionRequest:
    properties:
      enabled:
        type: boolean
      metric_type:
        type: string
    required:
    - metric_type
    type: object
  internal_group.TransferOwnershipRequest:
    properties:
      new_owner_id:
        type: string
    required:
    - new_owner_id
    type: object
  internal_group.UpdateGroupRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  internal_group.UpdateMemberStatusRequest:
    properties:
      status:
        enum:
        - accepted
        - rejected
        type: string
    required:
    - status
    type: object
  internal_group.UpdatePermissionsRequest:
    properties:
      metric_types:
        items:
          type: string
        type: array
    required:
    - metric_types
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: This is the API for the HealthMate application.
  title: HealthMate API
  version: "1.0"
paths:
...
  /groups:
    get:
      description: Get all groups (owned or joined) for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/auth-service_internal_domain.Group'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: List my groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Create a new group with the current user as owner
      parameters:
      - description: Group creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_group.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth-service_internal_domain.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new group
      tags:
      - groups
  /groups/{id}:
    delete:
      description: Remove a group (Owner only)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a group
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Update name and description of a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_group.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Update group info
      tags:
      - groups
  /groups/{id}/members:
    get:
      description: List all members of a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/auth-service_internal_domain.GroupMember'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Get group members
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Send an invitation to a user by email
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Invite details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_group.InviteMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Invite a member
      tags:
      - groups
  /groups/{id}/members/{member_id}:
    delete:
      description: Kick a member from the group (Owner only)
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Member User ID
        in: path
        name: member_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Remove a member
      tags:
      - groups
  /groups/{id}/members/me:
    delete:
      description: Leave a group you are a member of
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Leave a group
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Accept or reject a group invitation
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_group.UpdateMemberStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Update membership status
      tags:
      - groups
  /groups/{id}/owner:
    put:
      consumes:
      - application/json
      description: Transfer group ownership to a new user
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: New owner details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_group.TransferOwnershipRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Transfer ownership
      tags:
      - groups
  /groups/{id}/permissions:
    get:
      description: Get current user's sharing permissions in a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/auth-service_internal_domain.Permission'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Get my permissions
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Enable or disable sharing for a specific metric type
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Permission details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_group.SetPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Set sharing permission
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Update all sharing permissions for the current user
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: List of metric types
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_group.UpdatePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.OKResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth-service_internal_web_helpers.ErrorResponse'
      security:
      - Bearer: []
      summary: Update all permissions
      tags:
      - groups
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"